propagate load_sync_channel; reorganize testing directory
